fixed tests, split configuration into property and axon specific, re-enabled h2 to support JPA tests
