+++ core/datapool/datapool-core/src/main/kotlin/io/holunda/polyflow/datapool/core/business/DataEntryAggregate.kt
@@ -1,6 +1,7 @@
 package io.holunda.polyflow.datapool.core.business
 
 import io.holunda.camunda.taskpool.api.business.*
+import io.holunda.polyflow.datapool.core.DataPoolCoreAxonConfiguration
 import io.holunda.polyflow.datapool.core.DataPoolCoreConfiguration
 import io.holunda.polyflow.datapool.core.DeletionStrategy
 import mu.KLogging
@@ -17,9 +18,9 @@ import org.axonframework.spring.stereotype.Aggregate
  * The aggregate is manually created by the CreateOrUpdateCommandHandler.
  */
 @Aggregate(
-  repository = DataPoolCoreConfiguration.DATA_ENTRY_REPOSITORY,
-  snapshotTriggerDefinition = DataPoolCoreConfiguration.DATA_ENTRY_SNAPSHOTTER,
-  cache = DataPoolCoreConfiguration.DATA_ENTRY_CACHE,
+  repository = DataPoolCoreAxonConfiguration.DATA_ENTRY_REPOSITORY,
+  snapshotTriggerDefinition = DataPoolCoreAxonConfiguration.DATA_ENTRY_SNAPSHOTTER,
+  cache = DataPoolCoreAxonConfiguration.DATA_ENTRY_CACHE,
 )
 class DataEntryAggregate() {
