+++ core/datapool/datapool-core/src/test/kotlin/io/holunda/polyflow/datapool/core/DataPoolCoreDeletionStrategyConfigurationTest.kt
@@ -4,6 +4,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.axonframework.eventsourcing.Snapshotter
 import org.axonframework.eventsourcing.eventstore.EventStore
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertThrows
 import org.mockito.kotlin.mock
 import org.springframework.boot.context.annotation.UserConfigurations
 import org.springframework.boot.test.context.runner.ApplicationContextRunner
@@ -18,7 +19,8 @@ internal class DataPoolCoreDeletionStrategyConfigurationTest {
   fun `should do lax by default`() {
     contextRunner
       .withUserConfiguration(TestMockConfiguration::class.java)
-      .withPropertyValues(""
+      .withPropertyValues(
+        ""
       ).run {
         assertThat(it.getBean(DataPoolProperties::class.java)).isNotNull
         assertThat(it.getBean(DeletionStrategy::class.java)).isNotNull
@@ -64,6 +66,18 @@ internal class DataPoolCoreDeletionStrategyConfigurationTest {
 
         val deletionStrategy: DeletionStrategy = it.getBean(DeletionStrategy::class.java)
         assertThat(deletionStrategy.strictMode()).isFalse()
+      }
+  }
+
+  @Test
+  fun `should not allow wrong values in deletion mode`() {
+
+    contextRunner
+      .withUserConfiguration(TestMockConfiguration::class.java)
+      .withPropertyValues(
+        "polyflow.core.data-entry.deletion-strategy=hugo"
+      ).run {
+        assertThrows<IllegalStateException> { it.getBean(DataPoolProperties::class.java) }
       }
   }
