+++ core/datapool/datapool-core/src/test/kotlin/io/holunda/polyflow/datapool/core/itest/TestApplication.kt
@@ -1,4 +1,4 @@
-package io.holunda.polyflow.datapool
+package io.holunda.polyflow.datapool.core.itest
 
 import com.thoughtworks.xstream.XStream
 import com.thoughtworks.xstream.security.AnyTypePermission
