+++ integration/camunda-bpm/springboot-starter/src/test/kotlin/TaskpoolStarterITest.kt
@@ -1,3 +1,5 @@
+package io.holunda.polyflow.taskpool
+
 import io.holunda.polyflow.datapool.sender.DataEntryCommandSender
 import io.holunda.polyflow.taskpool.EnableTaskpoolEngineSupport
 import io.holunda.polyflow.taskpool.sender.task.EngineTaskCommandSender
