+++ view/jpa/pom.xml
@@ -51,6 +51,7 @@
     <dependency>
       <groupId>com.h2database</groupId>
       <artifactId>h2</artifactId>
+      <version>1.4.200</version> <!-- pin -->
       <scope>test</scope>
     </dependency>
     <dependency>
